Fold reset/retry tokens into the same module
