+++ quinn-proto/src/shared.rs
@@ -1,9 +1,9 @@
-use std::{cmp, fmt, net::SocketAddr, time::Instant};
+use std::{fmt, net::SocketAddr, time::Instant};
 
 use bytes::{Buf, BufMut, BytesMut};
 use rand::Rng;
 
-use crate::{coding::BufExt, packet::PartialDecode, MAX_CID_SIZE, RESET_TOKEN_SIZE};
+use crate::{coding::BufExt, packet::PartialDecode, ResetToken, MAX_CID_SIZE};
 
 /// Events sent from an Endpoint to a Connection
 #[derive(Debug)]
@@ -158,43 +158,6 @@ impl EcnCodepoint {
                 return None;
             }
         })
-    }
-}
-
-/// Stateless reset token
-///
-/// Used for an endpoint to securely communicate that it has lost state for a connection.
-#[allow(clippy::derive_hash_xor_eq)] // Custom PartialEq impl matches derived semantics
-#[derive(Debug, Copy, Clone, Hash)]
-pub struct ResetToken([u8; RESET_TOKEN_SIZE]);
-
-impl cmp::PartialEq for ResetToken {
-    fn eq(&self, other: &ResetToken) -> bool {
-        crate::constant_time::eq(&self.0, &other.0)
-    }
-}
-
-impl cmp::Eq for ResetToken {}
-
-impl From<[u8; RESET_TOKEN_SIZE]> for ResetToken {
-    fn from(x: [u8; RESET_TOKEN_SIZE]) -> Self {
-        Self(x)
-    }
-}
-
-impl std::ops::Deref for ResetToken {
-    type Target = [u8];
-    fn deref(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-impl fmt::Display for ResetToken {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        for byte in self.iter() {
-            write!(f, "{:02x}", byte)?;
-        }
-        Ok(())
     }
 }
